redeploy

+++ front-end/src/components/Preview.jsx
@@ -24,7 +24,7 @@ function Preview ({resumeInView}) {
                             {resumeInView.header_urls[headerindex+1] && "  |"}
                         </div>
                     })}
-                    
+
                 </div>
             </header>
             {resumeInView.resumesummary &&
@@ -193,7 +193,7 @@ function Preview ({resumeInView}) {
                                                     {"("}
                                                     {new Date(role.start).toLocaleDateString("en-IN", {
                                                         year: "numeric",
-                                                        month: "long"
+                                                        month: "short"
                                                     })}
                                                     {/* Show dash and end/ongoing only if start exists */}
                                                     {(role.end || role.ongoing) && (
@@ -201,7 +201,7 @@ function Preview ({resumeInView}) {
                                                             {" - "}
                                                             {/* Show end date if not ongoing and end exists */}
                                                             {!role.ongoing && role.end && new Date(role.end).toLocaleDateString("en-IN", {
-                                                                month: "long",
+                                                                month: "short",
                                                                 year: "numeric"
                                                             })}
                                                             {/* Show 'ongoing' only if ongoing is true */}
@@ -280,7 +280,7 @@ function Preview ({resumeInView}) {
                                     {subsection.start && (
                                         <>
                                         {new Date(subsection.start).toLocaleDateString("en-IN", {
-                                            month: "long",
+                                            month: "short",
                                             year: "numeric",
                                         })}
 
@@ -290,7 +290,7 @@ function Preview ({resumeInView}) {
                                             {/* Show end date if not ongoing and end exists */}
                                             {!subsection.ongoing && subsection.end &&
                                                 new Date(subsection.end).toLocaleDateString("en-IN", {
-                                                month: "long",
+                                                month: "short",
                                                 year: "numeric",
                                                 })}
 
@@ -393,7 +393,7 @@ function Preview ({resumeInView}) {
                                     {subsection.start && (
                                         <>
                                         {new Date(subsection.start).toLocaleDateString("en-IN", {
-                                            month: "long",
+                                            month: "short",
                                             year: "numeric",
                                         })}
 
@@ -403,7 +403,7 @@ function Preview ({resumeInView}) {
                                             {/* Show end date if not ongoing and end exists */}
                                             {!subsection.ongoing && subsection.end &&
                                                 new Date(subsection.end).toLocaleDateString("en-IN", {
-                                                month: "long",
+                                                month: "short",
                                                 year: "numeric",
                                                 })}
